paging inventory v1
